move gcs functions to base
